Merge short and full versions of "How to get the best out of this forum"
The "How to get the best out of this forum" stickies were originally published in a "short" and a "full" version. Most
people will be discouraged from learning the rules if they are presented with a large document. So the idea was for the
entry point to contain only the most essential information, while also providing a more detailed to use as a reference or
for those who require more information.

A large amount of the content was duplicated between the two documents, which greatly increases the effort required to
maintain the content and also makes it less likely that anyone would read both.

The Discourse forum software provides a "Hide Details" post formatting feature that allows text to be folded into a
clickable section of the post. This offers the possibility of achieving the original intent of the short vs. full approach
without the need to maintain two documents.

This should be a great improvement for the reader because it means they are only a single click away from the
detailed information on a specific subject of interest, rather than needing to scan through the "full" document to get to
it, perhaps only to discover that there was no additional information on that subject.