Add discussion topic links to moderator documentation topics
The "Moderation Guidelines" and "Moderator Instructions" topics should be used exclusively as a place to publish the
documentation. Discussion related to their content should be done in a separate topic. I have created a dedicated topic
for that purpose. Adding a link to the documentation topics will allow the discussion topic to be found easily.