Change "How to get the best out of this forum" image URLs to forum URLs
The previous image URLs were to the files in this repository, but those URLs can't be used for the content published on
the forum. The images were uploaded to the forum server when the topic was updated, and the URLs of the images in the
topic are to those copies of the image files. The source images will continue to be stored in this repository to allow
use in future development.