Update "How to get the best out of this forum" link in "FAQ/Guidelines" readme
The previous URL is for what was previously the "short version" topic, but that topic has since been replaced with a link
to what was previously the "full version" topic, which is now the sole copy of the content. It's more convenient for the
link to take you directly to the content, rather than needing to go through a layer of redirection.