Moderator Instructions: Add flags documentation
Flags are the mechanism used by Discourse to allow the automated systems or users to bring things to the attention of the
moderators, replacing SMF's "report to moderator" feature. The flags system is quite complex, so it's important to
provide moderators with good documentation of the system and how it is used.