Moderator Instructions: Add an overview for dealing with inappropriate behavior
One of the user-submitted flag types is "Inappropriate Behavior". The Discourse moderation system offers several tools to
deal with inappropriate behavior. There are separate instructions for each of those tools but it can be useful to provide
an overview of dealing with inappropriate behavior and the tools available to moderators.