Moderator Instructions: Specify that restriction of benevolent ads to Products and Services is for topics
Benevolent mentions of legitimate products and services may be completely appropriate in the course of discussion on any
topic in any category. However, topics dedicated solely to advertising things of interest to the Arduino community are
only allowed in the "Products and Services" category. These topics are considered spam in any other category.

The distinction between these two things was not clear from the previous wording.