How to get the best out of this forum: Use headings
Previously, the sections of the document were identified with bold text. With the markup provided by SMF, this was the
only real option available. Now we have more powerful markup. Use of headings for the document section titles allows the
heirarchical structure of the document that was already in place to be made more clear to the reader. This provides a
clear roadmap for the reader, reducing the chance of a "TL;DR" reaction.

In addition to being a guide a conscientious new forum user would read top to bottom, this document can also serve as a
useful reference. With that use case in mind, I have done two things:

Added HTML anchor tags to each heading.
Although anchors are automatically generated for each heading, Discourse has the infuriating behavior of appending an
index number to the ID of each anchor as a cheap way to avoid collisions. This means that all the auto-generated anchor
IDs change every time a new heading is added above. The workaround is to manually add anchors, which have custom IDs that
will never change.

Link headings to custom anchors.
Discourse automatically linkifies headings, which is super convenient because it means you can simply click on a heading
to get the URL to link directly to that heading, rather than needing to look at the source code for the anchor ID.
Unfortunately, the automated links point to the garbage auto-generated anchors. The workaround for this is to manually
linkify the heading text and point it to the custom permanent anchors.

The techniques mentioned above are used by Discourse themselves in the stock FAQ that comes with the installation:
https://forum.arduino.cc/t/faq-guidelines/5