Moderator Instructions: Adjust order of top level categories
Previously the order was:

- Flags
- General
- Guides

This seemed strange because The flags instructions are essentially guides, though the importance, quantity and related
documentation made it a candidate for a dedicated top level section. My thought is that the vast majority of the
moderation work is for a small set of specific occurrences, with standardized actions, so the "guide" instructions are
the most important, with the "general" instructions acting primarily as references for the "guide" content to link to.

For this reason, it makes more sense to put all the "guide" instructions at the top, starting with the "flags"
instructions, since flags are most often the triggers for moderation activity:

- Flags
- Guides
- General