Moderator Instructions: Correct anchor IDs
Although anchors are automatically generated for each heading, Discourse has the infuriating behavior of appending an
index number to the ID of each anchor as a cheap way to avoid collisions. This means that all the auto-generated anchor
IDs change every time a new heading is added above. It also causes the auto-generated table of contents to not work,
since the tool assumes that the reasonable, standard approach to anchor IDs is in use.

The workaround is to manually add anchors, which have custom IDs that will never change. The downside to this is that it
is a bit difficult to create and maintain the correct IDs manually. That results in the occasional screw-up, and thus the
corrections made here.