How to get the best out of this forum: Add sequence numbers to "Making your first post"
The "Making your first post" section of the document describes a process that must be done in an exact order. This
section is quite different in this respect from the rest of the content.

Prefixing the headings of this section with index numbers makes it clear how to use these instructions. I think it
becomes more approachable once the user sees that they only need to follow these four steps in order to accomplish
creating their first topic.