Moderator Instructions: Link to "Delete spam" instructions from other instructions related to spam
In addition to the instructions for directly deleting spam posts and accounts, the "Delete spam" section serves as the
primary area for documenting information related to spam. However, other instructions are also related to spam.
Integrating links to the primary spam documentation content from those sections makes the information more accessible to
the reader without any increase in verbosity.