Moderator Instructions: Link to documentation re: what is considered spam from the "Flags" section
In addition to the instructions for directly deleting spam posts and accounts, the "Delete spam" section serves as the
primary area for documenting information related to spam, including the specific exceptions where mentions of products or
services are not considered spam. However, the instructions under the "Flags" section are also directly related to
dealing with content reported as spam. For this reason, it's important to also clearly document what is considered spam
there. Rather than duplicating the content, doubling the maintenance burden, it's better to just link to the centralized
source of the information.