Canned Replies: Import from forum
This is the exact content of the Canned Replies as they exist on Discourse at this moment.

Each reply is stored in a dedicated folder. This facilitates:

- A dedicated README.md file for documenting the usage of the reply.
- The capacity for storing reply-specific images in the repository.

The reply's "Title" is defined via the title of the readme.

The reply's "Content" is defined by the text in the Markdown file named after the title.