Moderator Instructions: Document post order resulting from topic merge
In SMF, the merged topics were interlaced according to chronology. In Discourse, the posts are appended to the end of the
thread of the target topic. We are accustomed to it making no difference which we choose as the the source and target
topics when doing a full merge in SMF, so it might be useful for moderators to understand that this decision can now
affect the clarity of the resulting thread.