Switch to Task-based infrastructure assets
Previously, the project infrastructure was purely GitHub Actions based. If a contributor wanted to run any validations
locally, they would have to study the workflows to understand the necessary dependencies and commands.

Adding tasks for all common developer actions, then using those tasks in the GitHub Actions workflows, allows
contributors to more easily run the same operation locally as will be done by the CI system.

These are the state of the art assets sourced from the the centralized collection of reusable assets:

https://github.com/arduino/tooling-project-assets