Disable internal anchor checks by link checker tool
Validation of links to internal anchors was recently added to the markdown-link-check tool used to check for broken
links in this project's documentation.

Although a much needed feature, unfortunately it does not currently support anchors created by HTML anchor tags.

For example, this is valid and common Markdown:

[click here](#some-anchor)
<a name="some-anchor"></a>

but will fail the check:

ERROR: 1 dead links found!
[x] #some-anchor -> Status: 404

This type of link markup is used in the forum asset posts, which caused a spurious failure of the link check.

The solution is to configure markdown-link-check to skip these links (which was the behavior anyway with versions 3.8.7
and older). That will be reverted whenever the tool is able to correctly validate internal anchor links.