Add a project contributor guide
Documentation of how to contribute to the project gives everyone the opportunity to participate, while also reducing
maintenance effort and increasing quality of contributions.

This guide documents the various ways of contributing to the project.

It takes advantage of GitHub's "contributing guidelines" feature to increase the visibility of the information to
prospective contributors. The promoted "CONTRIBUTING.md" file is used as an entry point to the guide, with links from
there leading to dedicated guides for each specific type of contribution. In this way, the contributor is only presented
with relevant information, improving the approachability and readability of the content.