Add issue chooser and forms
A menu of issue categories will be presented when the user begins the issue creation process. This "template chooser"
also contains links to relevant information, which can redirect support requests or other inappropriate/unproductive
uses of the issue tracker.

Issues are submitted via custom web forms which ensure the necessary information is provided in a standardized forma.