Remove "Loop-Back Test Instructions" content
In order to enable collaborative development and maintenance into perpetuity, ownership of any content hosted in this
repository must be transferred to the forum's `@system` user.

The "Loop-Back Test Instructions" content has not been transferred, so storing it in this repository serves no purpose
and may waste the time of the contributors.