Mark non-public links
Some of the content is stored in categories of the forum not publicly visible (either due to only being of interest to
staff, or else previously public categories that are mothballed due to being obsolete).

Contributors without staff privileges, or even those with privileges but not signed in might be confused to click one of
these links and get a 404 if the nature of the target was not communicated.