Add Discourse contributor guide contact link to issue template chooser
Although the repository is used as an issue tracker for problems with the Arduino forum software, the scope is limited
to the configuration specific to the Arduino Forum. Defects or enhancements to the underlying Discourse platform must be
submitted to the upstream project.

The "issue template chooser" page is shown to the contributor when they initiate an issue creation process. Adding a
"contact link" to the Discourse contributor guide in this page will redirect contributors to the appropriate channel for
submitting those reports.