Add category order data
The forum software is configured to show categories in a specific order in the category listings. Defining the order in
this repository allows making efficient proposals for changes to the category order.

The order data is stored in a single file. The alternative would be to define the position value in the individual
category metadata, but that approach would require edits to multiple files when proposing a change to the order, which
is inconvenient and error-prone.