Add category permissions data
The "See" (view category and topics), "Reply" (reply to topics), and "Create" (create topics) permissions of a category
can be set for each of the forum's user groups.

Adding this data to the assets repository allows contributors to efficiently make proposals for adjustments of the
permissions (e.g., preventing creation of topics in a category that is intended to only serve as a "container" for
sub-categories).

The permissions of each category are shown in a table in the "**Permissions**" section of the `README.md` file under the
category folders.