Require trust level >1 to create topics in tutorials category
Unlike the other categories where all relevant topics are welcome, the tutorials forum category is a curated collection
of high quality tutorials contributed by the forum community.

New users regularly create topics requesting assistance in this category. These make work for the forum maintainers who
must move them to an appropriate category and make the category less useful as a source of tutorials.

The Discourse forum framework used by Arduino Forum has a user "trust level" system based on a user's forum activity. As
a user makes appropriate use of the forum, their trust level is automatically promoted and they gain more privileges.

In order to prevent miscategorized topics in this category, the configuration is changed to require a "trust level" of 2
or higher to directly create new topics in the tutorials category. Users with lower trust levels are still welcome to
propose the addition of their tutorials. A pinned topic is added to the category with instructions for doing this.

Since the change to the configuration is expected to prevent miscategorized topics in the category, the existing topics
warning against doing so are unpinned.