Add "Using Arduino > Uncategorized" category
It is common for new users to create topics in the "Using Arduino > Installation & Troubleshooting" category even though
the subject matter is not in scope for the intended usage of that category.

The hypothesis is that one factor in these miscategorizations is that the category happens to be the first available one
in the category listings and new users (and possibly spammers as well) are selecting it out of laziness instead of
taking the time to select the most appropriate category.

The "Uncategorized" category is added in that problematic first writable category position to act as a dedicated
collection point for these topics. The forum maintainers can periodically review the topics that are created there and
clear out the category.

Since the fact that the category should not be used is clearly communicated to the user, creation of a topic there is a
clear indication that the forum members should guide the user in the responsible use of the forum.