Print instructions to log if ToC is outdated
Some assets contain a table of contents to facilitate the navigation of their content.

The table of contents must be kept in sync with the content. The continuous integration system uses the excellent
markdown-toc tool to detect outdated table of contents.

In order to support checking any number of files at arbitrary paths in the repository, using a specified `--maxdepth`
flag value for each, a job matrix is used in the "Check ToC" GitHub Actions workflow. The user may also run the
`markdown:toc` locally to update the table of contents, but because the project-specific configuration is implemented in
the workflow instead of the taskfile they must specify the file's path and `--maxdepth` flag value via task variables in
the invocation. Contributors may find it challenging to update the table of contents.

In order to facilitate the update following detection of an out of sync state by the continuous integration system,
instructions for two resolution options are printed to the workflow run logs:

- exact command to run locally
- replacement with the corrected file from the workflow artifact