Add instructions re: "flag_sockpuppets" flag to Moderator Instructions
A new feature for automated identification of potentially malicious posts was recently enabled:

> If a new user replies to a topic from the same IP address as the user who started the topic, flag both of their posts
> as potential spam.

This system produce a new type of automated flag for the moderators to review. Flag review instructions are hereby added
for this flag type alongside the existing instructions for other automated flags in the "Moderator Instructions" asset
post.