Wrap standalone admonitions in horizontal rules
A common convention in technical writing is the "admonition", which adds supplemental information to the surrounding
normal prose.

These "admonitions" typically have a standardized styling that differentiates them from the rest of the prose in a
visually prominent manner. This often includes putting a "box" around the text (either a border or a background color).

Markdown doesn't have a dedicated admonition capability, but a standardized admonition style can be created using the
general purpose markup. I have settled on the use of a leading and trailing horizontal rule (markup: `---`) to create
something like a "box" around the admonition text.