Bold non-prose references in "Moderator Instructions"
Technical writing such as the "Moderator Instructions" asset post is often littered with non-prose text such as
references to:

- UI elements
- navigation items
- product names

The reader may have difficulty understanding when text is prose vs a reference to one of these items. For this reason,
readability is improved by formatting the text so that there is a clear visual difference between the standard prose
text and the non-prose items.

This can be accomplished by making these types of non-prose text bold.

That formatting convention was already employed in the document, but was missing in some places. The formatting is
hereby done consistently throughout the document.