Don't bold quoting of non-prose references in "Moderator Instructions"
Technical writing such as the "Moderator Instructions" asset post is often littered with non-prose text such as
references to:

- UI elements
- navigation items
- product names

The reader may have difficulty understanding when text is prose vs a reference to one of these items. For this reason,
readability is improved by formatting the text so that there is a clear visual difference between the standard prose
text and the non-prose items.

This is done by employing several formatting techniques:

- Bolding
- Quoting

Since the goal is only to emphasize the non-prose references, the quotes should not be bolded, so the nesting of the
formatting markup must be done in the correct order.

Some of the formatting was previously nested in the incorrect order, which resulted in inconsistent formatting and
harmed readability.