Use consistent formatting of `<kbd>` tags in Moderator Instructions
HTML kbd tags are used to create graphical representations of the forum user interface buttons in the "Moderator
Instructions" asset post. The visual appearance of the rendered content is enhanced by bolding text in the tags.

The nesting of the markup can generally be in either order. However, nesting the bolding markup inside the kbd tags
gives more flexibility by allowing text to be selectively bolded. For example, some graphical characters are clearer or
faithful to the UI when not bolded. So the standardized style is to nest the bolding markup inside the kbd tags.

Some of the markup did not comply with this standardized style.