Add line breaks to improve readability of lists in "Moderator Instructions"
The "Moderator Instructions" asset post consists largely of instructions in the form of ordered lists.

The first line of the step should contain only the essential information. Supplemental information must be placed on
subsequent lines so that it can easily be skipped over by those who are already knowledgeable on the subject.