Use standardized symbols to indicate nature of admonitions in "Moderator Instructions"
One of the elements used in the content of the "Moderator Instructions" asset post is "admonitions". These contain
information related to the surrounding content, formatted in a manner that makes them visually distinctive.

"Admonitions" fall into three general categories:

- Supplements: Additional information to help the reader to understand more about the subject
- Notices: Important information the reader should be aware of
- Warnings: Something bad can happen if an instruction is not performed correctly

In order to provide a clear visual indication of the type of an admonition, it is prefixed by a standardized symbolic
character.