Replace emoji markup with characters in "Moderator Instructions"
The "Moderator Instructions" asset post frequently refers to elements of the the forum user interface. That interface
makes heavy use of purely graphical elements in addition to text. Where possible, representations of those graphics are
used in the UI element references.

Previously, these mostly consisted of the colon bracketed emoji markup supported by the forum's post renderer. This
markup is an extension rather than standard part of the Markdown markup language so the support and rendering of the
markup may vary from one tool or platform to another. For example, the VS Code Markdown preview does not support the
markup, so the emojis are not rendered at all for contributors using that tool during the development of the asset.

This situation is improved by replacing the non-standard markup with the equivalent character. In addition to making the
content more universally comprehensible, this also provides slightly better control over these graphical characters (the
rendering of the emoji tends to change more dramatically over time).