Remove redundant line break tags from "Moderator Instructions"
The "Moderator Instructions" asset post contains a set of instructions in the form of an ordered list with each step
illustrated using a screenshot of the forum UI.

In order to improve readability, a blank line was inserted between the text of the step and the accompanying screenshot.
This can be accomplished in two ways:

- A `<br />` HTML tag
- A blank line (Markdown paragraph markup)

For some reason, both of the above were used. Although the redundant markup didn't affect the rendered content, it adds
unnecessary complexity to the source markup, making it less friendly to contributors and more difficult to maintain. For
this reason, the redundant markup is removed. Since it is the more intuitive of the two, the blank line markup was
retained.