Add infrastructure for checking sync of category order metadata
In order to allow contributors to include the specific order of the category listings in proposals, the repository
contains a metadata file that defines the category order.

This file is automatically generated using an API call.

A task and GitHub Actions workflow are added for the maintenance of this data.

On any change to a relevant file, and periodically, the workflow will run and compare the data in the repository against
the actual category structure of the forum to ensure the data remains in sync.

If the category order metadata file is found to be out of sync, the updated file is uploaded to a workflow artifact and
a message notifying the contributor of its availability printed to the workflow run logs. This will assist contributors
in easily bringing the file into sync.