Add taskfile to "Check ToC" workflow's paths filter
The repository's GitHub Actions workflows are configured to only run on changes to relevant files. This makes the
continuous integration system more efficient, but care must be taken to include all relevant files in the filter.

The "Check ToC" workflow uses a task to run the table of contents generation commands.

Previously the taskfile that defined that task was not included in the workflow's paths filter even though it is a
critical asset of the workflow.