Remove "Agree there is a problem" staff post template
A collection of precomposed replies for use in common moderation operations is provided to staff via the "Insert
template" menu of the post composer.

While the text of the "Agree there is a problem" template is appropriate when the forum software automatically adds it
as a reply to the flag review thread after a moderator makes a "Yes" review, there isn't any common occasion where a
staff member would need to manually create a reply with this text. So the template only cluttered up the menu.

For this reason, the template is hereby removed.