Remove "Code tags Usage" staff post template
A collection of precomposed replies for use in common moderation operations is provided to staff via the "Insert
template" menu of the post composer.

Replies for use in providing project support are out of scope for the templates since it is not feasible to develop and
maintain such content as part of the forum assets.

Since it is out of scope, the "Code tags Usage" template is removed.