Remove "We are only volunteers" staff post template
A collection of precomposed replies for use in standard moderation operations is provided to staff via the "Insert
template" menu of the post composer.

There isn't any standard moderation operation that would require a reply of this nature. So the template only cluttered
up the menu.

The template is hereby removed as superfluous.