Add note re: availability of "staging site" forum to "About the TEST category" post
While the "TEST" category is a convenient place for staff to make simple tests and experiments, it is insufficient for
some things. In this case, the dedicated sandbox forum instance can be used for freely running tests and experiments on
any aspect of the forum.

For this reason, it will be helpful to bring the existence of the "staging site" forum to the attention of the users of
the TEST forum. This is done by adding a note to the "About the TEST category" post, which links to instructions in a
topic dedicated to the subject.