Add "staging site" forum access instructions link to "Moderator Instructions" asset docs
The "Moderator Instructions" asset post contains a detailed instructions for each of the standard moderation procedures.

The procedures must be tested during the development and maintenance of the content. The "staging site" sandbox forum
instance is useful for this purpose; allowing the performance of the procedures freely without any risk of affecting the
production forum.

For this reason, it will be helpful to bring the existence of the "staging site" forum to the attention of those doing
development and maintenance of the "Moderator Instructions" content. This is done by adding a link to a topic dedicated
to the subject in the "Related" section of the "Moderator Instructions" asset's documentation.