Rename "Installation & Troubleshooting" category to "IDE 1.x"
The previous name of this category was very ambiguous. It could be interpreted by the user to be the appropriate
category for requesting assistance with solving problems with their project, even though we have categories specifically
dedicated to that purpose.

Based on the the term "Installation" in the category name and the original description:

> For problems with Arduino itself, NOT your project

the forum maintainers have interpreted the purpose of the category to be discussion related to Arduino IDE (which is
sometimes referred to as "Arduino").

The maintainers have restricted the use of the category to that subject matter for years and the current category
description clearly explains the scope. However, a disproportionate number of miscategorized topics are still created in
the category. The previous ambiguous category name is likely a significant factor in this misuse.

We now have a dedicated category for discussion of Arduino IDE 2.x, so it is necessary to specify that this category is
only for Arduino IDE 1.x-related discussion.

The naming follows the established convention of dropping the "Arduino" from the software product name from the category
where that is implicit (e.g., "Web Editor", "IoT Cloud", "PLC IDE").