Rename "Software" category to "Development Tools"
In the world of embedded systems development, the term "software" might be used to refer to several distinct things.
Although an experienced forum user might be able to infer from the subcategories that the "software" referred to by the
category name is the applications used for development and deployment of Arduino projects rather than the
software/firmware that make up the projects themselves, such an inference should not be required and the ambiguity may
make it difficult for a less experienced user to understand the scope of the category and its subcategories.

The name "Development Tools" is more explicit while also remaining reasonably concise. A category description is added
in the "About the Development Tools category" post (which has a special technical treatment by the forum software) in
order to even more clearly define the scope of the category.