Rename "Arduino IDE 2.0" category to "IDE 2.x"
A long term existence was not given consideration at the time of the naming of the "Arduino IDE 2.0" category. The name
has already been made outdated by the release of Arduino IDE 2.1.0.

The new name follows the established convention of dropping the "Arduino" from the software (and hardware) product name
from the category name where that is implicit (e.g., "Web Editor", "IoT Cloud", "PLC IDE").