Rename "Programming Questions" category to "Programming"
The scope of this category is discussion of subjects related to programming.

The "questions" of the previous category name is implicit. Category names should be kept to the minimum necessary to
effectively communicate the purpose of the category (with the description used to provide details about the purpose and
scope).