Merge away "India" category
The category is essentially unused (2 legitimate topics created there in the past year).

Unlike the other categories under the "International" parent category, the topics under the "India" category are almost
entirely written in English. So it doesn't serve the special purpose of collecting discussions in a specific non-English
language.

Those who create topics there receive less assistance than they would have if they posted in the more actively monitored
categories instead.

The harm that results from leading users to post in a place that is contrary to their best interest, as well as the
increased complexity added to the forum category structure greatly outweighs the benefits of the category, so it is
hereby removed.

The topics in the category have been moved individually to the appropriate remaining categories.

Topic and post URLs are not category-dependent, so this change will not impact any links to topics or posts. A redirect
has been set up from the category URL to the removal announcement post, which explains the reason for the category's
removal and that the general English language categories should be used instead.