Merge away "India" category
The category is essentially unused (2 legitimate topics created there in the past year).

Unlike the other categories under the "International" parent category, the topics under the "India" category are almost
entirely written in English. So it doesn't serve the special purpose of collecting discussions in a specific non-English
language.

Those who create topics there receive less assistance than they would have if they posted in the more actively monitored
categories instead.

The harm that results from leading users to post in a place that is contrary to their best interest, as well as the
increased complexity added to the forum category structure greatly outweighs the benefits of the category, so it is
hereby removed.

The topics in the category will be moved individually to the appropriate remaining categories.