Delete "Garbage" category
This category is an artifact of the less capable forum framework used years ago. At that time, the category was used as
a way to remove posts from public visibility while still allowing them to be restored in the event a removal was done in
error.

The category is no longer used because the Discourse forum framework uses a different system for post deletion (which
still allows recovery).

So the category serves no purpose. It has the harmful effect of making the forum more confusing for new staff members
and clutters up the forum category structure for all staff members. For this reason, it is hereby deleted. The posts it
contains will be deleted, but they can still be viewed by staff members when investigating an account that is suspected
of malicious behavior.