Update Arduino Science Journal App reference link
The previous link currently produces a 503 response. That may be a temporary situation, but during my investigation I
discovered that the previous link has been a redirect for some time now. Experience shows that these redirects have a
tendency to break or change to leading to content other than what was intended (e.g., to the arduino.cc home page). So
even if the redirect may eventually be fixed, it is better to update the link to point directly to the intended content
rather than relying on the redirect.