Add category ID 191 to order check skip list
The repository's "Check Category Order Metadata" GitHub Actions workflow compares the metadata in
`content/categories/order.md` against the actual category structure on the forum in order to check for missed updates to
order.md as well as for incorrect category structure on the forum.

A well designed category structure is important for making the forum user friendly. However, in the case of vestigial or
internal categories with limited access, the structure is not important and so there is no need to include them in this
check. In this case, the category ID is added to the array in the skip list.