Use appropriate slug for Hardware > Portenta > Edge Control category
Discourse adds an optional path component (referred to as the "slug" to forum category URLs to make them more human
readable. This "slug" is automatically generated from the category name by default, but the admin can set a custom slug
if they like.

For unknown reasons that "slug" was set to "edge-impulse" at the time of the creation of the category, even though that
is inappropriate for the stated subject matter of the category.

The slug is hereby changed to the appropriate default "edge-control".

This won't have any impact on existing links to the category because Discourse disregards the value of this path
component entirely, using instead the machine identifier category ID component of the URL (178), which is not affected
by the change, and redirecting any URL containing that ID to the appropriate
https://forum.arduino.cc/c/hardware/portenta/edge-control/178 URL regardless of what the slug component is.