Configure link checker tool to ignore Chrome App link
The markdown-link-check tool is used to detect broken links in the project's Markdown files.

The tool is detecting the link to the Arduino Create Chrome App page in the Chrome Web Store as broken:

ERROR: 1 dead links found in ./content/categories/software/chrome-app/_topics/about-the-chrome-app-category/1.md !
[✖] https://chromewebstore.google.com/detail/arduino-create-for-educat/elmgohdonjdampbcgefphnlchgocpaij → Status: 0
task: Failed to run task "markdown:check-links": exit status 1

However, when the link is visited by a human it works as expected. This spurious failure is avoided by adding that URL
to the tool's ignore list.