Add moderator instructions for "Illegal" flags
The Moderator Instructions document defines the standard procedures moderators follow to handle flags raised on posts.
When a user raises a flag, they select from a list of general reasons for flagging. The Moderator Instructions document
provides a procedure for each of these flag types.

Since the time the Moderator Instructions document was written, Discourse added a new "Illegal" reason to the flag
dialog. A procedure for handling this flag type is hereby added.