Add link to flagging instructions in "FAQ/Guidelines" asset post
The reader won't necessarily know what is meant by "flag", or how to do it. Adding a detailed explanation and
instructions to the post would have a harmful effect, but linking the mention of flagging to the documentation provided
by Discourse makes the information readily available to those readers who need it, while not impacting those who don't.